refactoring code with routes
