+++ package.json
@@ -9,7 +9,9 @@
     "cors": "^2.8.4",
     "express": "^4.16.3",
     "express-session": "^1.15.6",
+    "helmet": "^3.13.0",
     "knex": "^0.15.2",
+    "morgan": "^1.9.1",
     "sqlite3": "^4.0.2"
   },
   "devDependencies": {
